# backend/cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/kpi/presence-rate/{user_uuid}/{start_date}/{end_date}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the presence rate percentage for a specified user UUID between the provided start and end dates. 🔒 Requires role: **any**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KPI"
                ],
                "summary": "Get presence rate for a user within a date range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start Date in ISO 8601 format",
                        "name": "start_date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End Date in ISO 8601 format",
                        "name": "end_date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "user_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.KPIPresenceRateResponse"
                        }
                    }
                }
            }
        },
        "/kpi/work-session-team-weekly-total/{team_uuid}/{start_date}/{end_date}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the total work session time in minutes for a specified team UUID between the provided start and end dates. 🔒 Requires role: **manager**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KPI"
                ],
                "summary": "Get total work session time for a team within a date range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start Date in ISO 8601 format",
                        "name": "start_date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End Date in ISO 8601 format",
                        "name": "end_date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Team UUID",
                        "name": "team_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.KPIWorkSessionTeamWeeklyTotalResponse"
                        }
                    }
                }
            }
        },
        "/kpi/work-session-user-weekly-total/{user_uuid}/{start_date}/{end_date}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the total work session time in minutes for a specified user UUID between the provided start and end dates. 🔒 Requires role: **any**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KPI"
                ],
                "summary": "Get total work session time for a user within a date range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start Date in ISO 8601 format",
                        "name": "start_date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End Date in ISO 8601 format",
                        "name": "end_date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "user_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.KPIWorkSessionUserWeeklyTotalResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.KPIPresenceRateResponse": {
            "type": "object",
            "properties": {
                "presence_rate": {
                    "type": "number"
                },
                "user_uuid": {
                    "type": "string"
                },
                "weekly_rate_expected": {
                    "type": "number"
                },
                "weekly_time_done": {
                    "type": "number"
                }
            }
        },
        "model.KPIWorkSessionTeamMemberWeeklyTotal": {
            "type": "object",
            "properties": {
                "total_time": {
                    "type": "integer"
                },
                "user_uuid": {
                    "type": "string"
                }
            }
        },
        "model.KPIWorkSessionTeamWeeklyTotalResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.KPIWorkSessionTeamMemberWeeklyTotal"
                    }
                },
                "start_date": {
                    "type": "string"
                },
                "team_uuid": {
                    "type": "string"
                },
                "total_time": {
                    "type": "integer"
                }
            }
        },
        "model.KPIWorkSessionUserWeeklyTotalResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "total_time": {
                    "type": "integer"
                },
                "user_uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}